prevent adding empty task , used toastr for alerting while adding TODO , used sweetalert2 for confirming while completing the TODO | deleting the TODO

+++ js/main.js
@@ -27,7 +27,6 @@
  *        :   "custom headers" : ...
  *        :   "cach-control" : "no-cache"
 */
-
 const formElement = document.querySelector("form");
 const inputElement = document.querySelector("input");
 const apiKey = "6949582f600ccb137ffe54be";
@@ -40,7 +39,9 @@ formElement.addEventListener('submit', (e) => { //submit = enter | press button[
     e.preventDefault();
     // console.log("hi");
     // console.log(inputElement.value);
-    addTODO();
+    if (inputElement.value.trim().length > 0) {
+        addTODO();
+    }
 })
 // =========================================================================
 async function addTODO() {
@@ -71,9 +72,15 @@ async function addTODO() {
         if (data.message === "success") {
 
             //getAllTODOs : [displayAllTODOs]
+            toastr.success('Added Successfuly :)', "Add a new TODO")
             await getAllTODOs(); //it's asyn   => so i need await to avoid to make the reset of the form done before it finish
             // console.log(data);//{message : "success"}
             formElement.reset();//will not reset untill [allTODOs return]
+        }
+        else {// still it's not an error from the server , but backender called me ^^
+            // console.log(data);//{message:'error' , error:[message:...]}
+            // console.log(data.error[0].message);//the msg from backend[server] =>  title is not allowed to be empty"
+            toastr.error(`${data.error[0].message}`, 'Error happend!')
         }
 
     }
@@ -114,7 +121,7 @@ function displayAllTODOs() {
         cartona += `
 
         <li class="d-flex justify-content-between align-items-center border-bottom pb-2 my-2">
-                <span onclick="markCompleted('${todoElement._id}');" class="task-name ${todoElement.completed ? `completed` : ``}" style="${todoElement.completed ? `text-decoration: line-through;` : ``}">${todoElement.title}</span>
+                <span ${todoElement.completed ? `` :`onclick="markCompleted('${todoElement._id}');"` } class="task-name ${todoElement.completed ? `completed` : ``}" style="${todoElement.completed ? `text-decoration: line-through;` : ``}">${todoElement.title}</span>
                 <div class="d-flex align-items-center gap-4 ">
                     <span><i class="fa-solid fa-circle-check ${todoElement.completed ? `d-block` : `d-none`}" style="color: #63E6BE;"></i></span>
                     <span onclick="deleteTODO('${todoElement._id}')" class="trash-icon-wrapper rounded"><i class="fa-solid fa-trash text-white "></i></span>
@@ -132,32 +139,50 @@ function displayAllTODOs() {
 
 async function markCompleted(todoID) {
 
-    const todoData = {
-        todoId: todoID
-    }
-    const obj = {
-        method: "put", // put : for update
-        body: JSON.stringify(todoData),
-        headers: {
-            "content-type": "application/json"
+    Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "rgb(90, 83, 189)",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, make it complete!"
+    }).then(async (result) => {
+        if (result.isConfirmed) {
+
+            const todoData = {
+                todoId: todoID
+            }
+            const obj = {
+                method: "put", // put : for update
+                body: JSON.stringify(todoData),
+                headers: {
+                    "content-type": "application/json"
+                }
+            }
+
+
+
+            const response = await fetch("https://todos.routemisr.com/api/v1/todos", obj);
+            // console.log(response);
+            if (response.ok) {
+                const data = await response.json();
+                // console.log(data);
+                if (data.message === "success") {// in backend ->the todo become {... , completed = true , ...}
+                    //will not call displayAllTODOs() ...because it use allTODOs array which is empty after loading ^^
+                    //so will use getAllTODOs.. contain displayAllTODOs and update the array with the data from the API ^^
+                    getAllTODOs();
+                    Swal.fire({
+                        title: "Completed :)",
+                        icon: "success"
+                    });
+
+
+                }
+            }
+
         }
-    }
-
-
-
-    const response = await fetch("https://todos.routemisr.com/api/v1/todos", obj);
-    // console.log(response);
-    if (response.ok) {
-        const data = await response.json();
-        // console.log(data);
-        if (data.message === "success") {// in backend ->the todo become {... , completed = true , ...}
-            //will not call displayAllTODOs() ...because it use allTODOs array which is empty after loading ^^
-            //so will use getAllTODOs.. contain displayAllTODOs and update the array with the data from the API ^^
-            getAllTODOs();
-
-
-        }
-    }
+    });
 
 
 }
@@ -166,27 +191,55 @@ async function markCompleted(todoID) {
 async function deleteTODO(id) {
 
 
+    Swal.fire({
+        title: "Are you sure?",
+        text: "You won't be able to revert this!",
+        icon: "warning",
+        showCancelButton: true,
+        //   confirmButtonColor: "#3085d6",
+        confirmButtonColor: "rgb(90, 83, 189)",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Yes, delete it!"
+    }).then(async (result) => {
+        if (result.isConfirmed) {
 
-    const todoData = {
-        todoId: id
-    };
-    const obj = {
-        method: "delete",
-        body: JSON.stringify(todoData),
-        headers: {
-            "content-type": "application/json"
-        }
-    };
+            //begin the logic of delete after conforming first :)
+            const todoData = {
+                todoId: id
+            };
+            const obj = {
+                method: "delete",
+                body: JSON.stringify(todoData),
+                headers: {
+                    "content-type": "application/json"
+                }
+            };
 
-    const response = await fetch("https://todos.routemisr.com/api/v1/todos", obj);
-    if(response.ok){
-        const data = await response.json();
-        if(data.message === "success"){
-            getAllTODOs();
+            const response = await fetch("https://todos.routemisr.com/api/v1/todos", obj);
+            if (response.ok) {
+                const data = await response.json();
+                if (data.message === "success") {
+                    Swal.fire({ // event will fire 
+                        title: "Deleted!",
+                        icon: "success"
+                    });
+                    getAllTODOs();
+                }
+            }
+
+
+
         }
-    }
+    });
+
+
+
+
+
+
+
+
 
 
 }
 // ====================================
-
